+++ nucleoid/lib/src/html/tag_element.dart
@@ -240,3 +240,5 @@ abstract class TimeBuilt implements BuiltNodeElement<ChildNodeElement> {
     Attribute(name: 'datetime'),
   ];
 }
+
+abstract class CodeBuilt implements BuiltNodeElement<ChildNodeElement> {}