+++ nucleoid/CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.1.0
+## 0.1.1
+- Add class `Route`
+```
+$ pub run nucleoid create your_project_name
+```
 
+## 0.1.0
 - Initial version